refactor(tests): compact and parametrization tests. fix non-cleanable world (purge after run)
