+++ module.txt
@@ -1,6 +1,6 @@
 {
     "id" : "FlexibleMovement",
-    "version" : "0.2.0-SNAPSHOT",
+    "version" : "0.3.0-SNAPSHOT",
     "author" : "",
     "displayName" : "FlexibleMovement",
     "description" : "",
